+++ WeatherForecast/WeatherForecast/WeatherTableViewCell.swift
@@ -44,7 +44,11 @@ class WeatherTableViewCell: UITableViewCell {
         let dashLabel: UILabel = UILabel()
         descriptionLabel = UILabel()
         
-        let labels: [UILabel] = [dateLabel, temperatureLabel, weatherLabel, dashLabel, descriptionLabel]
+        let labels: [UILabel] = [dateLabel, 
+                                 temperatureLabel,
+                                 weatherLabel,
+                                 dashLabel,
+                                 descriptionLabel]
         
         labels.forEach { label in
             label.textColor = .black
